+++ slskit/commands.py
@@ -1,3 +1,4 @@
+import difflib
 import sys
 from pathlib import Path
 from unittest.mock import patch
@@ -35,14 +36,33 @@ def refresh(config: Config):
 
 
 def create_snapshot(config: Config):
-    highstate = state.show_highstate(config)
-    snapshot = {minion_id: result.value for minion_id, result in highstate.items()}
-    dump = salt.utils.yaml.safe_dump(snapshot, default_flow_style=False)
+    dump = _dump_highstate(config)
     config.snapshot_path.write_text(dump)
 
 
 def check_snapshot(config: Config):
-    print("check_snapshot")
+    if not config.snapshot_path.exists():
+        sys.exit(f"Snapshot file {config.snapshot_path} not found")
+
+    snapshot = config.snapshot_path.read_text()
+    current = _dump_highstate(config)
+
+    if snapshot != current:
+        _display_diff(snapshot, current)
+        sys.exit(1)
+
+
+def _dump_highstate(config: Config) -> str:
+    highstate = state.show_highstate(config)
+    snapshot = {minion_id: result.value for minion_id, result in highstate.items()}
+    return salt.utils.yaml.safe_dump(snapshot, default_flow_style=False)
+
+
+def _display_diff(a: str, b: str):
+    diff = difflib.unified_diff(
+        a.splitlines(keepends=True), b.splitlines(keepends=True)
+    )
+    sys.stdout.writelines(diff)
 
 
 def _display_output(output: dict, config: Config):